+++ member-service/member-domain/member-application-service/src/main/java/com/project/imdang/member/service/domain/handler/auth/KaKaoApiClientHandler.java
@@ -1,32 +1,30 @@
-package com.project.imdang.member.service.domain.handler;
+package com.project.imdang.member.service.domain.handler.auth;
 
 import com.project.imdang.member.service.domain.dto.oauth.OAuthLoginCommand;
 import com.project.imdang.member.service.domain.dto.oauth.OAuthLoginResponse;
-import com.project.imdang.member.service.domain.dto.oauth.google.GoogleLoginResponse;
-import com.project.imdang.member.service.domain.dto.oauth.google.GoogleTokenResponse;
+import com.project.imdang.member.service.domain.dto.oauth.kakao.KakaoLoginResponse;
 import com.project.imdang.member.service.domain.valueobject.OAuthType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
+import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 
-
 @Component
 @RequiredArgsConstructor
-public class GoogleApiClientHandler implements OAuthApiClientHandler {
+public class KaKaoApiClientHandler implements OAuthApiClientHandler{
     private final RestTemplate restTemplate;
 
-    @Value("${oauth.google.url.api}")
+    @Value("${oauth.kakao.url.api}")
     private String apiUrl;
 
     @Override
     public OAuthType oAuthType() {
-        return OAuthType.GOOGLE;
+        return OAuthType.KAKAO;
     }
 
     @Override
@@ -38,10 +36,13 @@ public class GoogleApiClientHandler implements OAuthApiClientHandler {
     @Override
     public OAuthLoginResponse getOAuthInfo(String accessToken) {
         HttpHeaders httpHeaders = new HttpHeaders();
-        httpHeaders.setContentType(MediaType.APPLICATION_JSON);
-        httpHeaders.add(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken);
+        httpHeaders.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
+        httpHeaders.set("Authorization", "Bearer " + accessToken);
 
-        HttpEntity<?> request = new HttpEntity<>(httpHeaders);
-        return restTemplate.exchange(apiUrl, HttpMethod.GET, request, GoogleLoginResponse.class).getBody();
+        MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
+        body.add("property_keys", "[\"kakao_account.profile]\"]");
+
+        HttpEntity<?> request = new HttpEntity<>(body, httpHeaders);
+        return restTemplate.postForObject(apiUrl, request, KakaoLoginResponse.class);
     }
 }
