+++ member-service/member-domain/member-application-service/src/main/java/com/project/imdang/member/service/domain/handler/auth/OAuthLoginCommandHandler.java
@@ -1,4 +1,4 @@
-package com.project.imdang.member.service.domain.handler;
+package com.project.imdang.member.service.domain.handler.auth;
 
 import com.project.imdang.member.service.domain.MemberDomainService;
 import com.project.imdang.member.service.domain.dto.LoginResponse;
