+++ member-service/member-domain/member-application-service/src/main/java/com/project/imdang/member/service/domain/handler/auth/TokenRequestHandler.java
@@ -1,4 +1,4 @@
-package com.project.imdang.member.service.domain.handler;
+package com.project.imdang.member.service.domain.handler.auth;
 
 import com.project.imdang.domain.jwt.JwtTokenProvider;
 import com.project.imdang.member.service.domain.dto.TokenResponse;
