+++ member-service/member-domain/member-application-service/src/main/java/com/project/imdang/member/service/domain/handler/member/DetailMyPageCommandHandler.java
@@ -1,4 +1,4 @@
-package com.project.imdang.member.service.domain.handler;
+package com.project.imdang.member.service.domain.handler.member;
 
 import com.project.imdang.domain.valueobject.MemberId;
 import com.project.imdang.member.service.domain.dto.DetailMyPageQuery;
